Corrected an issue in xxdr.c that I introduced, changed how hdf5 is located by cmake.

+++ oc2/xxdr.c
@@ -220,7 +220,7 @@ xxdr_free(XXDR* xdr)
 int
 xxdr_skip(XXDR* xdrs, off_t len)
 {
-  ssize_t pos;
+    unsigned int pos;
     pos = xxdr_getpos(xdrs);
     pos = (pos + len);
     if(pos < 0) pos = 0;
